requirements.txt
adal==1.2.6
amqp==5.0.5
asgiref==3.3.1
attrs==20.3.0
autopep8==1.5.5
bandit==1.7.0
billiard==3.6.3.0
cachetools==4.2.1
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cligj==0.7.1
colorama==0.4.4
coolprop==6.4.1
cryptography==3.4.6
Cython==0.29.21
decorator==4.4.2
Django==3.1.8
django-celery-results==2.0.1
django-extensions==3.1.1
django-picklefield==3.0.1
django-tastypie==0.14.3
docopt==0.6.2
et-xmlfile==1.0.1
Fiona==1.8.18
future==0.18.2
geopandas==0.6.3
gitdb==4.0.5
GitPython==3.1.13
google-api-core==1.26.0
google-api-python-client==1.12.8
google-auth==1.27.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
greenlet==1.0.0
gunicorn==20.0.4
h5pyd==0.8.0
httplib2==0.19.0
idna==2.10
importlib-metadata==3.4.0
ipython-genutils==0.2.0
isodate==0.6.0
jdcal==1.4.1
kombu==5.0.2
lxml==4.6.3
more-itertools==8.7.0
msrest==0.6.21
msrestazure==0.6.4
munch==2.5.0
numpy==1.17.5
oauthlib==3.1.0
openpyxl==3.0.6
packaging==20.9
pandas==0.25.3
pathlib2==2.2.1
pbr==5.5.1
pexpect==4.2.1
pipreqs==0.4.10
prompt-toolkit==3.0.16
protobuf==3.15.0
psutil==5.6.7
psycopg2==2.7.3.2
ptyprocess==0.5.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.7.4
PyJWT==2.0.1
pyparsing==2.4.7
pyproj==1.9.5.1
python-dateutil==2.8.1
python-mimeparse==1.5.5
pytz==2019.3
PyYAML==5.4
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
rollbar==0.14.7
rsa==4.7.1
Shapely==1.7.1
six==1.15.0
smmap==3.0.5
SQLAlchemy==1.3.23
sqlparse==0.4.1
stevedore==3.3.0
toml==0.10.2
traitlets==4.3.3
typing-extensions==3.7.4.3
tzlocal==1.5.1
uritemplate==3.0.1
urllib3==1.26.5
vine==5.0.0
wcwidth==0.1.9
xlrd==1.0.0
yarg==0.1.9
zipp==0.6.0
